+++ src/Config.hpp
@@ -91,6 +91,8 @@ public:
   bool enableIntraBC = true;
   bool enableAngleDelta = true;
   bool enableDiagonalIntra = true;
+  bool enableDirectionalIntra = true;
+  bool enableTxSizeSearch = true;
   bool lossless = false;
   bool enableCDEF = false;
   bool enableRestoration = false;